+++ tests/integration/conftest.py
@@ -33,11 +33,8 @@ from pytest_operator.plugin import OpsTest
 from state import (
     BASE_IMAGE_CONFIG_NAME,
     BUILD_INTERVAL_CONFIG_NAME,
-    DOCKERHUB_CACHE_CONFIG_NAME,
     EXTERNAL_BUILD_FLAVOR_CONFIG_NAME,
     EXTERNAL_BUILD_NETWORK_CONFIG_NAME,
-    JUJU_CHANNELS_CONFIG_NAME,
-    MICROK8S_CHANNELS_CONFIG_NAME,
     OPENSTACK_AUTH_URL_CONFIG_NAME,
     OPENSTACK_PASSWORD_CONFIG_NAME,
     OPENSTACK_PROJECT_CONFIG_NAME,
@@ -215,12 +212,6 @@ def openstack_connection_fixture(clouds_yaml_contents: str) -> Connection:
     return openstack.connect(first_cloud)
 
 
-@pytest.fixture(scope="module", name="dockerhub_mirror")
-def dockerhub_mirror_fixture(pytestconfig: pytest.Config) -> str:
-    """Dockerhub mirror URL."""
-    return pytestconfig.getoption("--dockerhub-mirror") or ""
-
-
 @pytest.fixture(scope="module", name="test_id")
 def test_id_fixture() -> str:
     """The test ID fixture."""
@@ -229,7 +220,10 @@ def test_id_fixture() -> str:
 
 @pytest.fixture(scope="module", name="test_configs")
 def test_configs_fixture(
-    model: Model, charm_file: str, test_id: str, dispatch_time: datetime, dockerhub_mirror: str
+    model: Model,
+    charm_file: str,
+    test_id: str,
+    dispatch_time: datetime,
 ) -> TestConfigs:
     """The test configuration values."""
     return TestConfigs(
@@ -237,7 +231,6 @@ def test_configs_fixture(
         charm_file=charm_file,
         dispatch_time=dispatch_time,
         test_id=test_id,
-        dockerhub_mirror=dockerhub_mirror,
     )
 
 
@@ -246,14 +239,11 @@ def image_configs_fixture():
     """The image configuration values used to parametrize image build."""
     return ImageConfigs(
         bases=("noble",),
-        juju_channels=tuple(),  # ("3.5/stable",), juju support will be removed
-        microk8s_channels=tuple(),  # ("1.29-strict/stable",), microk8s support will be removed
     )
 
 
 @pytest.fixture(scope="module", name="app_config")
 def app_config_fixture(
-    test_configs: TestConfigs,
     private_endpoint_configs: PrivateEndpointConfigs,
     image_configs: ImageConfigs,
     openstack_metadata: OpenstackMeta,
@@ -262,9 +252,6 @@ def app_config_fixture(
     return {
         BASE_IMAGE_CONFIG_NAME: ",".join(image_configs.bases),
         BUILD_INTERVAL_CONFIG_NAME: 12,
-        DOCKERHUB_CACHE_CONFIG_NAME: test_configs.dockerhub_mirror,
-        JUJU_CHANNELS_CONFIG_NAME: ",".join(image_configs.juju_channels),
-        MICROK8S_CHANNELS_CONFIG_NAME: ",".join(image_configs.microk8s_channels),
         REVISION_HISTORY_LIMIT_CONFIG_NAME: 5,
         OPENSTACK_AUTH_URL_CONFIG_NAME: private_endpoint_configs["auth_url"],
         OPENSTACK_PASSWORD_CONFIG_NAME: private_endpoint_configs["password"],
@@ -440,14 +427,6 @@ def image_names_fixture(image_configs: ImageConfigs, app: Application):
     arch = _get_supported_arch()
     for base in image_configs.bases:
         image_names.append(f"{app.name}-{base}-{arch.value}")
-        for juju in image_configs.juju_channels:
-            for microk8s in image_configs.microk8s_channels:
-                image_names.append(
-                    (
-                        f"{app.name}-{base}-{arch.value}-juju-{juju.replace('/','-')}"
-                        f"-mk8s-{microk8s.replace('/','-')}"
-                    )
-                )
     return image_names
 
 
@@ -471,58 +450,4 @@ async def bare_image_id_fixture(
         check_interval=30,
     )
     assert image, "Bare image not found"
-    return image.id
-
-
-@pytest_asyncio.fixture(scope="module", name="juju_image_id")
-async def juju_image_id_fixture(
-    openstack_connection: Connection,
-    dispatch_time: datetime,
-    image_configs: ImageConfigs,
-    app: Application,
-):
-    """The Juju bootstrapped image expected from builder application."""
-    arch = _get_supported_arch()
-    image: Image | None = await wait_for(
-        functools.partial(
-            image_created_from_dispatch,
-            image_name=(
-                f"{app.name}-{image_configs.bases[0]}-{arch.value}-juju-"
-                f"{image_configs.juju_channels[0].replace('/','-')}-"
-                f"mk8s-{image_configs.microk8s_channels[0].replace('/','-')}"
-            ),
-            connection=openstack_connection,
-            dispatch_time=dispatch_time,
-        ),
-        timeout=60 * 30,
-        check_interval=30,
-    )
-    assert image, "Juju image not found"
-    return image.id
-
-
-@pytest_asyncio.fixture(scope="module", name="microk8s_image_id")
-async def microk8s_image_id_fixture(
-    openstack_connection: Connection,
-    dispatch_time: datetime,
-    image_configs: ImageConfigs,
-    app: Application,
-):
-    """The Juju bootstrapped image expected from builder application."""
-    arch = _get_supported_arch()
-    image: Image | None = await wait_for(
-        functools.partial(
-            image_created_from_dispatch,
-            image_name=(
-                f"{app.name}-{image_configs.bases[0]}-{arch.value}-juju-"
-                f"{image_configs.juju_channels[0].replace('/','-')}-"
-                f"mk8s-{image_configs.microk8s_channels[0].replace('/','-')}"
-            ),
-            connection=openstack_connection,
-            dispatch_time=dispatch_time,
-        ),
-        timeout=60 * 30,
-        check_interval=30,
-    )
-    assert image, "Microk8s image not found"
     return image.id